backend/requirements.txt
asgiref==3.11.1
certifi==2026.2.25
charset-normalizer==3.4.7
Django==6.0.3
django-cors-headers==4.9.0
django-stubs==5.2.9
django-stubs-ext==5.2.9
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
gunicorn==25.3.0
idna==3.11
librt==0.8.1
mypy==1.20.0
mypy_extensions==1.1.0
packaging==26.0
pathspec==1.0.4
PyJWT==2.12.0
python-dotenv==1.2.2
requests==2.33.1
sqlparse==0.5.5
types-PyYAML==6.0.12.20250915
typing_extensions==4.15.0
urllib3==2.6.3
